implement ugly-number-ii

+++ src/games/tree/ugly-number-ii/Play.tsx
@@ -1,113 +1,79 @@
-import { HeapItem, useAlgoContext } from "./AlgoContext";
-import { Button, ButtonGroup, Stack, Table, TableBody, TableCell, TableRow, Typography } from '@mui/material';
+import { useAlgoContext } from "./AlgoContext";
+import { Button, ButtonGroup, Stack, Typography } from '@mui/material';
 import { wait } from "../../../data-structures/_commons/utils";
 import { State } from "./AlgoState";
 import SortIcon from '@mui/icons-material/Sort';
 import PlayCircleFilledWhiteOutlinedIcon from '@mui/icons-material/PlayCircleFilledWhiteOutlined';
+import { styled } from '@mui/system';
+
+const DisplayNumber = styled(Typography)({
+    border: "1px solid #bdbdbd",
+    borderRadius: "50%",
+    height: 45,
+    width: 45,
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "center",
+});
 
 const DisplayInput = () => {
-    const { current, matrix, k, completed } = useAlgoContext();
-
-    const inComplete = (row: number, col: number): boolean => {
-        for (let i = 0; i < completed.length; i++) {
-            const position = completed[i];
-            if (position.row === row && position.col === col) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    const getBackgroundColor = (row: number, col: number) => {
-        if (inComplete(row, col)) {
-            return "#bdbdbd";
-        }
-        if (row === current.row && col === current.col) {
-            return "green";
-        }
-        return "#fff";
-    }
-
-    const getColor = (row: number, col: number) => {
-        if (inComplete(row, col)) {
-            return "#fff";
-        }
-        if (row === current.row && col === current.col) {
-            return "#fff";
-        }
-        return "#000";
-    }
-
+    const { n } = useAlgoContext();
     return (
         <div style={{
             position: "fixed",
             top: "25%",
             left: "16%",
         }}>
-            <Stack direction="row" spacing={1} sx={{ display: "flex", alignItems: "center", justifyContent: "center" }}>
-                <Typography
-                    variant="h6"
-                    sx={{
-                        border: "1px solid #bdbdbd",
-                        borderRadius: "50%",
-                        height: 45,
-                        width: 45,
-                        display: "flex",
-                        alignItems: "center",
-                        justifyContent: "center",
-                    }}
-                >
-                    K
-                </Typography>
-                <Typography
-                    variant="h6"
-                    sx={{
-                        border: "1px solid #bdbdbd",
-                        borderRadius: "50%",
-                        height: 45,
-                        width: 45,
-                        display: "flex",
-                        alignItems: "center",
-                        justifyContent: "center",
-                    }}
-                >
-                    {k}
-                </Typography>
+            <Stack direction="row" spacing={1} sx={{ display: "flex", alignItems: "center", justifyContent: "center", color: "green" }}>
+                <DisplayNumber variant="h6">N</DisplayNumber>
+                <DisplayNumber variant="h6">{n}</DisplayNumber>
             </Stack>
+        </div>
+    );
+};
 
-            <div style={{ textAlign: "center", marginTop: "30px" }}>
-                <Typography variant="h6">Matrix</Typography>
-            </div>
+const factors = [2, 3, 5];
 
-            <Table>
-                <TableBody>
+const DisplayFactors = () => {
+    const { factorIndex, tip } = useAlgoContext();
+
+    const getVariant = (i: number) => {
+        if (factorIndex === undefined) {
+            return "outlined";
+        }
+        if (factorIndex === i) {
+            return "contained";
+        }
+        return "outlined"
+    }
+
+    return (
+        <div style={{
+            position: "fixed",
+            top: "35%",
+            left: "16%",
+        }}
+        >
+            <Stack direction="column" spacing={2} justifyContent="center">
+                <Typography color="green" variant="h5">
+                    {tip}
+                </Typography>
+                <ButtonGroup size="large" color="success">
                     {
-                        matrix.map((row, i) =>
-                            <TableRow key={i}>
-                                {
-                                    row.map((col, j) =>
-                                        <TableCell
-                                            key={j}
-                                            padding="none"
-                                            sx={{
-                                                color: getColor(i, j),
-                                                width: "65px",
-                                                textAlign: "center",
-                                                paddingTop: 0.4,
-                                                paddingBottom: 0.4,
-                                                backgroundColor: getBackgroundColor(i, j)
-                                            }}
-                                        >
-                                            <Typography variant="h6">
-                                                {col}
-                                            </Typography>
-                                        </TableCell>)
-                                }
-                            </TableRow>
+                        factors.map((factor, i) =>
+                            <Button
+                                key={i}
+                                sx={{ width: 45, height: 45 }}
+                                variant={getVariant(i)}
+                            >
+                                <Typography variant="h5">
+                                    {factor}
+                                </Typography>
+                            </Button>
                         )
                     }
-                </TableBody>
-            </Table>
+                </ButtonGroup>
+            </Stack>
         </div>
     );
 }
@@ -116,7 +82,7 @@ const DisplayResult = () => {
     const { result } = useAlgoContext();
 
     const Content = () => (
-        <Button size="large" variant="contained" color="success">
+        <Button size="large" variant="outlined" color="success">
             <Typography variant="h5">
                 {result}
             </Typography>
@@ -137,20 +103,14 @@ const DisplayResult = () => {
 
 const Main = () => {
 
-    const { animate, cancelAnimate, state, setState, minHeap, matrix, setCurrent, completed, setResult, setK, k } = useAlgoContext();
+    const { animate, cancelAnimate, state, setState, minHeap, setResult, n, setN, seen, setFactorIndex, setTip } = useAlgoContext();
 
     const handleBuildHeap = async () => {
         if (!minHeap) return;
         setState(State.Computing);
         animate();
         try {
-            for (let row = 0; row < matrix.length; row++) {
-                const col = 0;
-                const item = new HeapItem(matrix[row][col], row, col);
-                setCurrent({ row, col });
-                await minHeap.insert(item);
-            }
-            setCurrent({ row: 0, col: 0 });
+            await minHeap.insert(1);
             await wait(0.1);
         } catch (error) {
             console.error(error);
@@ -161,45 +121,41 @@ const Main = () => {
 
     const handleRun = async () => {
         if (!minHeap) return;
+        if (!seen) return;
         setState(State.Computing);
         animate();
         try {
             const root = await minHeap.delete();
             if (root) {
-                const { val, row, col } = root;
-                if (col + 1 < matrix[row].length) {
-                    const item = new HeapItem(matrix[row][col + 1], row, col + 1);
-                    setCurrent({ row: item.row, col: item.col });
-                    await minHeap.insert(item);
-                }
-                completed.push({ row, col });
-                setResult(val);
-
-                if (k !== 1) {
-                    const peek = await minHeap.peek();
-                    if (peek) {
-                        const { row, col } = peek;
-                        setCurrent({ row, col });
+                setResult(root);
+                for (let i = 0; i < factors.length; i++) {
+                    const factor = factors[i];
+                    const next = root * factor;
+                    setFactorIndex(i);
+                    setTip(`${root} * ${factor} = ${next}`);
+                    if (!seen.has(next)) {
+                        await minHeap.insert(next);
+                        seen.add(next);
                     }
-                    setState(State.Playing);
-                } else {
-                    setState(State.Finished);
-                    setCurrent({ row, col });
                 }
             }
             await wait(0.1);
         } catch (error) {
             console.error(error);
         }
+        setFactorIndex(undefined);
+        setTip(undefined);
         cancelAnimate();
-        setK(k => k - 1);
+        setState(State.Playing);
+        setN(n => (n !== undefined) ? n - 1 : undefined);
     }
 
     return (
         <>
-            {(matrix.length || "") && <DisplayInput />}
-
+            {(n !== undefined) && <DisplayInput />}
+            <DisplayFactors />
             <DisplayResult />
+
             <div style={{
                 position: "fixed",
                 bottom: "150px",
@@ -220,7 +176,7 @@ const Main = () => {
                         color="success"
                         startIcon={<PlayCircleFilledWhiteOutlinedIcon />}
                         onClick={handleRun}
-                        disabled={state !== State.Playing || k === 0}
+                        disabled={state !== State.Playing || n === 0}
                     >
                         run
                     </Button>
